Management_Software/requirements.txt
beautifulsoup4==4.8.1
bootstrap4==0.1.0
Django==2.2.7
django-bootstrap4==1.0.1
django-phone-field==1.8.0
pytz==2019.3
soupsieve==1.9.5
sqlparse==0.3.0
